requirements.txt
boto3==1.28.73
celery==5.3.4
cryptography==41.0.7
django-autoslug==1.9.9
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-storages==1.14.2
Django==4.2.10
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
gunicorn==21.2.0
hiredis==2.2.3
psycopg2-binary==2.9.9
redis==5.0.1
whitenoise==6.6.0